Fix server tests and improve code quality of tests.
Instead of passing around the name of the database, the backend passes around a MongoDatabase.
Only one MongoClient is initialized per class.
All tests now use the data-for-testing-only database (!!)
All mock databases are now called testDB
